Significant speedups through the use of numba
Using numba to JIT-compile some of the code speeds up seam carving on a
1280x720 image from about 2s/seam to 0.2s-0.25s/seam - an 8x-10x
improvement. Unfortunately, numba doesn't seem to play nice will all the
functions I've written, so the gradient_magnitude function specifically
is not JIT'd at this time.

Initially I tried to get speedups by avoiding some seemingly unnecessary
copies, but it proved to be a pessimism from about 2s/seam to 4s/seam.
Boo.

After a bit of research the most common thing in regards to speedups
that I kept seeing come up was parallelization, but it always came with
the caveat that it's very difficult to implement and may not be too
significant due to synchronization. So that's off the table for now.

Not gonna lie, a little disappointed that speedups came from effectively
changing the programming language itself rather than some "real"
optimization of my own doing, but it was easy performance I guess? And
this is cool to know about and will have to be kept in mind for future
projects.

Maybe there's still room for some optimization anyway. :-)

numba came to my attention through this blog post from Karith Karanth,
which itself came to my attention from a Github repo for another python
seam carving tool by Andrew Campbell.

https://karthikkaranth.me/blog/implementing-seam-carving-with-python/
https://github.com/andrewdcampbell/seam-carving